requirements.txt
flask
flask-cors
flask-session
flask-jsonpify
pyyaml
requests
pyjwt
cryptography
elasticsearch>=1.0.0,<2.0.0
os-package-registry>=0.0.12
blinker>=1.1
